Rfid hotfix

See merge request IAN/mxbi-project/mxbi!40

+++ src/mxbi/detector/detector_factory.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
 class DorsetLID665v42Config:
     baudrate: int
     port: str
-    interval: int
+    interval: float
 
 
 DetectorConfig = Union[DorsetLID665v42Config, None]
